requirements.txt
Flask>=2.0
pyTelegramBotAPI>=4.0
python-dotenv
Werkzeug